requirements.txt
Django==4.2
djangorestframework==3.14.0
drf_yasg==1.21.5
openai==0.27.4
pinecone_client==2.2.1
psycopg2==2.9.6
scipy==1.10.1
termcolor==2.3.0
textract==1.6.5
tiktoken==0.3.3
